ENH dedicated modules for orange output formats
