+++ educe/rst_dt/learning/cmd/extract.py
@@ -9,38 +9,24 @@ Extract features to CSV files
 """
 
 from __future__ import print_function
-import codecs
-import csv
 import os
 import sys
 
 import educe.corpus
-import educe.learning.keys
 import educe.glozz
 import educe.stac
 import educe.util
 from ..args import add_usual_input_args
 from ..base import read_corpus_inputs, extract_pair_features
+from ....learning.orange_format import dump_orange_tab_file
+
 
 NAME = 'extract'
-
-
-def mk_csv_writer(keys, fstream):
-    """
-    start off csv writer for a given mode
-    """
-    csv_quoting = csv.QUOTE_MINIMAL
-    writer = educe.learning.keys.OrangeWriter(fstream,
-                                              keys,
-                                              quoting=csv_quoting)
-    writer.writeheader()
-    return writer
 
 
 # ----------------------------------------------------------------------
 # options
 # ----------------------------------------------------------------------
-
 
 def config_argparser(parser):
     """
@@ -70,87 +56,28 @@ def config_argparser(parser):
                              '(currently none)')
     parser.set_defaults(func=main)
 
+
 # ---------------------------------------------------------------------
 # main
 # ---------------------------------------------------------------------
 
-
-def main_parsing_pairs(args):
-    """
-    Main to call when live data are passed in (--parsing). Live data are data
-    that we want to discourse parsing on, so we don't know if they are attached
-    or what the label is.
-
-    There used to be an expectation that live data was also flat data
-    (given with --live), but as of 2014-03-24, we are experimenting with
-    have hierarchical live data
-    """
-    feature_set = args.feature_set
-    inputs = read_corpus_inputs(args)
-    features_file = os.path.join(args.output, 'extracted-features.tab')
-    with codecs.open(features_file, 'wb') as ofile:
-        header = feature_set.PairKeys(inputs)
-        writer = mk_csv_writer(header, ofile)
-        feats = extract_pair_features(inputs,
-                                      feature_set=feature_set,
-                                      live=True)
-        for row, _ in feats:
-            writer.writerow(row)
-
-
-def _write_pairs(gen, r_ofile, p_ofile):
-    """
-    Given a generator of pairs and the relations/pairs output
-    file handles:
-
-    * use first row as header, then write the first row
-    * write the rest of the rows
-
-    If there are no rows, this will throw an StopIteration
-    exception
-    """
-    # first row
-    p_row0, r_row0 = gen.next()
-    p_writer = mk_csv_writer(p_row0, p_ofile)
-    r_writer = mk_csv_writer(r_row0, r_ofile)
-    p_writer.writerow(p_row0)
-    r_writer.writerow(r_row0)
-    # now the rest of them
-    for p_row, r_row in gen:
-        p_writer.writerow(p_row)
-        r_writer.writerow(r_row)
-
-
-def main_corpus_pairs(args):
-    """
-    The usual main. Extract feature vectors from the corpus
-    """
-    feature_set = args.feature_set
-    inputs = read_corpus_inputs(args)
-    of_bn = os.path.join(args.output, os.path.basename(args.corpus))
-    of_ext = '.tab'
-    if not os.path.exists(args.output):
-        os.makedirs(args.output)
-
-    relations_file = of_bn + '.relations' + of_ext
-    edu_pairs_file = of_bn + '.edu-pairs' + of_ext
-    with codecs.open(relations_file, 'wb') as r_ofile:
-        with codecs.open(edu_pairs_file, 'wb') as p_ofile:
-            gen = extract_pair_features(inputs, feature_set=feature_set)
-            try:
-                _write_pairs(gen, r_ofile, p_ofile)
-            except StopIteration:
-                # FIXME: I have a nagging feeling that we should properly
-                # support this by just printing a CSV header and nothing
-                # else, but I'm trying to minimise code paths and for now
-                # failing in this corner case feels like a lesser evil :-/
-                sys.exit("No features to extract!")
-
-
 def main(args):
     "main for feature extraction mode"
+    # retrieve parameters
+    inputs = read_corpus_inputs(args)
+    feature_set = args.feature_set
+    live = args.parsing
 
-    if args.parsing:
-        main_parsing_pairs(args)
+    # extract instances
+    X = extract_pair_features(inputs, feature_set=feature_set, live=live)
+    
+    # dump instances to file
+    if not os.path.exists(args.output):
+        os.makedirs(args.output)
+    if live:
+        out_file = os.path.join(args.output, 'extracted-features.tab')
     else:
-        main_corpus_pairs(args)
+        of_bn = os.path.join(args.output, os.path.basename(args.corpus))
+        of_ext = '.tab'
+        out_file = '{}.relations{}'.format(of_bn, of_ext)
+    dump_orange_tab_file(X, [], out_file)
