+++ educe/stac/learning/cmd/extract.py
@@ -16,7 +16,7 @@ import sys
 
 from educe.stac.learning import features
 import educe.corpus
-import educe.learning.keys
+from educe.learning.csv_format import KeyGroupWriter
 import educe.glozz
 import educe.stac
 import educe.util
@@ -29,9 +29,7 @@ def mk_csv_writer(keys, fstream):
     start off csv writer for a given mode
     """
     csv_quoting = csv.QUOTE_MINIMAL
-    writer = educe.learning.keys.KeyGroupWriter(fstream,
-                                                keys,
-                                                quoting=csv_quoting)
+    writer = KeyGroupWriter(fstream, keys, quoting=csv_quoting)
     writer.writeheader()
     return writer
